requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
stripe==7.10.0
beautifulsoup4==4.12.3
pytest==7.4.4
pytest-django==4.7.0
openai>=1.0.0
python-dateutil>=2.8.2
